Pulled out clients into a new file
Made the connection behaviour a separate file and the client modules
much more lightweight.
Switched from using send / receive calls to using :oneway and :request calls
